Privacy policy
The privacy policy for my app
Ikotun oluwafemi built this app as a free app. This service is provide by Ikotun oluwafemi at all cost and is intended for use as is. 
This page is used to inform visitor regarding my policy with the collection, use and disclosure of personal information if anyone decided to use my service.
If you choose to use my service, the you agree to the collection and use of information in relation to the policy. The personal information that i collect is used for providing and improving the service. I will not use or share your information with anyone.
The term used in this privacy policy have the same meaning as our term and conditions, which is accessible she'd plans unless otherwise definited in the policy. 
Information collection and use for a better experience while using our service, I may require you to provide us certain personal identifiable information. The information that I request will be returned on your device and is not collected by me in any away. 
The app does use third party service that may collect information use to identify you link to privacy policy of third party service provider used by the app 
Google play store
   Log Data
I want to inform you that whenever you use my service, in a case of an error in the app I collect data and information on your phone called Log Data
This Logo Data may include information such as your device internet's protocol (IP) address configuration of the app when utilizing my service, the time and date of you use of the service and other statics 
 Cookies  
Cookies are files with a small amount of data that are comonly use as anonymous unique identifier. These are sent to your browser from the website that you visit and are stored on your phone or internal memory 
This service deos not use those "cookies" explicitly. However, the app may use third party code and libraries that use "cookies" to collect information and improve. You have the option to either accept or refused these cookies, you may not be able to use some portion of the service 
Service provider
I may employ third party companies and individual due to the following reasoning